FCFS Algorithm seems to be working perfectly.  Now to finish FCFS.
To finish we have to add the summary printout.  We also need to determine if the program is meant to start running from cycle 0.  That is, should cycle 0 count as a cycle?  And in the verbose printout, are we saying that the number on the right is the number of time remaining for each particular activity?